Constrain Clang and LLD searches to LLVM version (#8634)
As of LLVM 17 or so, the *ConfigVersion.cmake files have been correctly
shipped with Clang and LLD, not just LLVM. The FindHalide_LLVM.cmake
module now searches for the exact matching versions of Clang and LLD.
Unfortunately, neither the LLVM_PACKAGE_VERSION variable (which is
documented as being safe to use with CMake version comparisons) nor the
LLVM_VERSION variable (which is undocumented but used internally) are
suitable for use with CMake version comparisons. LLVM's CMake build
continues to be a fractal of insanity. Instead, we reconstruct a valid
version string from the major/minor/patch versions.

Fixes #8633