Rework strict_float to use individual op intrinsics instead (#8641)
* Rework strict_float to use individual op intrinsics instead

This commit removes the strict_float intrinsic, which was problematic
because was a scoped thing, so hoisting an Expr out of it can lead to
incorrect codegen. This is hard for various lowering passes to handle.

Instead, this adds strict versions of all the floating point ops and
comparisons that either round or need to worry about nan/inf:
strict_add, strict_mul, strict_eq, etc. The strict_float helper is now a
mutator that replaces all adds, muls, eqs, etc with the strict intrinsic
equivalents.

The simplifier has no rules that handle these intrinsics, so it just
naturally doesn't do anything it shouldn't. The changes to the
simplifier look large, but they are all just removing
can-I-simplify-this checks, which changes indentation.

There is also a helper that does the reverse, which can be used by
analysis or codegen passes. The C-like backends just unstrictify these
intrinsics when encountered, and it's on the consumer to not set
fast-math flags. When CodeGen_LLVM encounters an intrinsic it sets a
scoped flag to indicate it is now in strict mode, and then recursively
codegens the unstrictified version of the op. If the strict_float target
flag is set, this scoped flag is just always true. CodeGen_LLVM defaults
every floating point operation to be strict. When it encounters an op
that has a strict equivalent but the scoped flag is currently false, it
temporarily sets the floating point flags to relaxed for that op only.
In this way it fails safe - use of fast math flags must be explicitly
requested at each emission of a floating point op.

Without this fail-safe behavior there were problems where something like
a vector slice operation was marked as no-nans, which was then fed into
a strict comparison, making the comparison not actually strict because
the tag on the previous op meant that LLVM was allowed to assume the
arguments were not nan. no-nans and similar are viral.

The above only applies if there is any usage of strict float intrinsics
in the module or if the target flag is set. If not, it just defaults all
ops to fast, as with main. For the purposes of this test, is_nan,
is_inf, and is_finite all count as strict float intrinsics.

There should be no visible changes to users. There are no performance
differences in the apps with the flag off.

* Address review comments

- Make unstrictify non-recursive, and evaluate args outside of the
strict scope in CodeGen
- Fix C++ emission and add preprocessor warning that looks for common
fast-math flags
- Remove strict_sqrt

* Add asserts

* Fix class vs struct for forward decls