Bounds and alignment analysis through bitwise ops (#8574)
* Analyze bitwise ops using a mask of bits known

converting it to and from the existing ExprInfo as necessary

* Bug fixes and better analysis of bitwise not

* Reorder test to know one more bit about narrow positive ints

* Slight simplification with no functional change

* Remove debugging print

* Propagate info in more places

Some fuzzing with extra checks turned on revealed that the simplifier
would often produce constants without the ExprInfo knowing they were
constants. In a few places we were doing a janky remutation of a known
constant just to set the info, but this wasn't being consistently done.
This commit changes constant-folding in the simplifier and info handling
in boolean ops to know that it's a constant value being returned.

Also standardize the spelling of zeros

* empty commit

* empty commit

* Minor tweaks

* Comment improvements

* Remove some may_simplify checks to make merge conflict easier to resolve