Get raw Runtime::Buffer from Buffer in Python rather than use PyBuffer
The buffer protocol (PyBuffer) does support all the same metadata that
Halide::Runtime::Buffer does. In particular, it does not support
coordinate offsets (as in set_min).

When an object is passed to an AOT Python pipeline, it is now checked
for a _get_raw_halide_runtime_buffer method. If it exists, it should
return a pointer to a halide_buffer_t whose lifetime is at least that
of the object.

Fixes #8652