Refactor bounds of a Call node (#8828)
This commit separates out the lowering of intrinsics to other
expressions to facilitate bounds inference from the actual handling of
call nodes. By getting the lowering code off the recursive path, it
reduces the size of the stack frame for the call visitor.

The actual bounds logic is intended to be unchanged.