Fix incorrect `argmax` rfactor pattern
The `solve_expression` call in Associativity.cpp moves `x`
values (the old values in the update definition) as far to
the left as possible. This prevents the `select()` call in
the existing argmax pattern from ever matching because it
compares `y0 < x0` rather than `x0 < y0`.

We fix this bug and add a test to prevent regressions.