Require disclosure of tool (e.g. LLM) usage to generate code (#8625)
* Require disclosure of tool (e.g. LLM) usage to generate code

Vibe coding is a hot topic. In our experience, it doesn't produce acceptably-good code right now, but it is steadily improving, so we probably need a policy on it. Following a discussion in the dev meeting, we feel that it's fine to use tools like LLMs to generate Halide PRs. They will be held to the same standard as any other PR. The human that opens the PR will be treated as responsible for the code. In fact, we already have a history of accepting tool-generated code (some simplifier rules were discovered by superoptimization). However, we decided contributors must disclose the tools used in any significantly-tool-generated code, for a few reasons. 1) You shouldn't claim credit for something you didn't do. 2) It affects how a reviewer will review the code (e.g. what kind of bugs they will look for). 3) In a variety of possible future scenarios, it may be useful to know which commits were substantially authored by these tools.

The credit issue makes it worthy of insertion into the code of conduct.

* Update CODE_OF_CONDUCT.md

Explicitly request "co-authored-by" tag