Improve `argmax` patterns for rfactor (#8863)
The `argmax` and `argmin` patterns in AssociativeOpsTable assumed that
comparisons would be normalized to `y < x` (or `x < y`), but the Solve
module could flip these. It was also possible an earlier pass would
negate the patterns such that `y >= x` (or `x >= y`) with the branches
reversed could be encountered instead.

To solve this issue, we extend Solve to move the variable of interest
to the `true` branch of a `select`. We also extend the table to cover
all four comparison cases.

We also substantially refactor Solve.cpp to use a helper to track and
update the ambient tree traversal state (`uses_var` and `failed`).