Add RuntimeLibraryInfo to pass pipeline for LLVM >= 22
https://github.com/llvm/llvm-project/pull/168622 adds a new analysis for
determining what libcalls are supported by a target. We need to add it
to the Halide pass pipeline, otherwise all libcalls are marked
unsupported. This results in some memory libcalls being lowered to loops
instead, which can increase runtime. This manifested for us as timeouts
for the web assembly tests due to the WABT interpreter taking
significantly longer to interpret memset loops versus the libcall.