Add ability to get the type of a FuncRef (#8895)
This is mostly relevant in the python bindings, where FuncRefs might get
passed around and treated as Exprs, including asking for their type. In
the C++ front-end such functions would typically take an Expr, but the
same problem could arise in generic code.

Also improved the error message if you try to call an undefined Func in
python. It was something about it not returning a Tuple. The general fix
seemed to be to add a better error message when casting a FuncRef of an
undefined Func to a Tuple.